♻️ refactor: 누누 코드리뷰 반영
