+++ nova-app/src/app/(pages)/layout.tsx
@@ -1,5 +1,9 @@
 const PublicLayout = ({ children }: { children: React.ReactNode }) => {
-  return <div className='h-dvh mx-auto bg-white-full'>{children}</div>;
+  return (
+    <div className='flex items-center justify-center min-h-screen mx-auto bg-white-full'>
+      {children}
+    </div>
+  );
 };
 
 export default PublicLayout;
