+++ nova-app/src/features/onboarding/models/useOnBoarding.tsx
@@ -1,22 +1,22 @@
 import { useState } from 'react';
-import { STEP_ITEMS } from '../data/StepItems';
-import { Step } from '../types/StepItem';
+import { STEP_ITEMS } from '@/features/onboarding/data/StepItems';
+import { Step } from '@/features/onboarding/types/StepItem';
 
-const STEPS = [1, 2, 3, 4] as const;
+const STEPS = ['step1', 'step2', 'step3', 'step4'] as const;
 
 export const useOnboarding = () => {
-  const [currentStep, setCurrentStep] = useState<Step>(1);
+  const [currentStep, setCurrentStep] = useState<Step>('step1');
   const [stepValidMap, setStepValidMap] = useState<Record<Step, boolean>>({
-    1: false,
-    2: false,
-    3: false,
-    4: true,
+    step1: false,
+    step2: false,
+    step3: false,
+    step4: true,
   });
 
   const isFirstStep = currentStep === STEPS[0];
   const isLastStep = currentStep === STEPS[STEPS.length - 1];
   const isStepValid = stepValidMap[currentStep];
-  const showSkip = currentStep === 2 || currentStep === 4;
+  const showSkip = currentStep === 'step2' || currentStep === 'step4';
 
   const onNext = () =>
     setCurrentStep((prev) => STEPS[Math.min(STEPS.indexOf(prev) + 1, STEPS.length - 1)]);
