+++ nova-app/src/features/onboarding/types/StepItem.ts
@@ -1,12 +1,13 @@
 import type { ComponentType } from 'react';
 
-export type Step = 1 | 2 | 3 | 4;
+export type Step = 'step1' | 'step2' | 'step3' | 'step4';
 
-export interface StepComponentProps {
+export type StepComponent = {
   onValidChange: (isValid: boolean) => void;
-}
+};
+
 export type StepItem = {
   label: string;
   description: string;
-  Component: ComponentType<StepComponentProps>;
+  Component: ComponentType<StepComponent>;
 };
