+++ nova-app/src/features/onboarding/ui/OnboardingContainer/MajorCard.tsx
@@ -1,11 +1,14 @@
 'use client';
 
-import { StepComponentProps } from '@/features/onboarding/types/StepItem';
 import { ToggleButton } from '@/shared/ui';
 import { useEffect, useState } from 'react';
-import { SKILL_OPTIONS } from '../data/SkillOptions';
+import { MAJOR_OPTIONS } from '@/features/onboarding/data/MajorOptions';
 
-const SkillCard = ({ onValidChange }: StepComponentProps) => {
+interface MajorCardProps {
+  onValidChange: (isValid: boolean) => void;
+}
+
+const MajorCard = ({ onValidChange }: MajorCardProps) => {
   const [selected, setSelected] = useState<string | null>(null);
 
   useEffect(() => {
@@ -18,7 +21,7 @@ const SkillCard = ({ onValidChange }: StepComponentProps) => {
 
   return (
     <div className='w-150 h-24.5 flex flex-wrap gap-2.5'>
-      {SKILL_OPTIONS.map((text) => (
+      {MAJOR_OPTIONS.map((text) => (
         <ToggleButton
           size='lg'
           key={text}
@@ -26,11 +29,11 @@ const SkillCard = ({ onValidChange }: StepComponentProps) => {
           variant='outline'
           selected={selected === text}
           onClick={() => toggleItem(text)}
-          className='w-73.5 h-11'
+          className='w-[142.5px] h-11'
         />
       ))}
     </div>
   );
 };
 
-export default SkillCard;
+export default MajorCard;
