+++ nova-app/src/features/onboarding/ui/Stepper.tsx
@@ -1,16 +1,20 @@
 import clsx from 'clsx';
+import { Step } from '@/features/onboarding/types/StepItem';
 
 interface StepperProps {
-  currentStep: number;
+  currentStep: Step;
   labels: string[];
 }
 
 const Stepper = ({ currentStep, labels }: StepperProps) => {
+  const STEP_ORDER: Step[] = ['step1', 'step2', 'step3', 'step4'];
+  const currentIndex = STEP_ORDER.indexOf(currentStep) + 1;
+
   return (
     <div className='flex items-center'>
-      {Array.from({ length: 4 }).map((_, index) => {
+      {STEP_ORDER.map((_, index) => {
         const step = index + 1;
-        const isActive = step <= currentStep;
+        const isActive = step <= currentIndex;
         const label = labels[index] ?? `Step ${step}`;
 
         return (
@@ -22,29 +26,28 @@ const Stepper = ({ currentStep, labels }: StepperProps) => {
                   'h-0.75 w-13.75 mr-2 rounded-r-static-pill',
                   step === 1
                     ? 'bg-transparent'
-                    : step <= currentStep
+                    : step <= currentIndex
                       ? 'bg-charcoal-full'
                       : 'bg-slate-outline',
                 )}
               />
 
-              <div className='flex flex-col items-center gap-1'>
-                <div
-                  className={clsx(
-                    'flex items-center justify-center w-6 h-6 rounded-static-pill typo-body-key',
-                    isActive ? 'bg-peak text-peak' : 'bg-surface text-charcoal-additive',
-                  )}
-                >
-                  {step}
-                </div>
+              <div
+                className={clsx(
+                  'flex items-center justify-center w-6 h-6 rounded-static-pill typo-body-key',
+                  isActive ? 'bg-peak text-peak' : 'bg-surface text-charcoal-additive',
+                )}
+              >
+                {step}
               </div>
 
+              {/* 오른쪽 선 */}
               <div
                 className={clsx(
                   'ml-2 h-0.75 w-13.75 rounded-l-static-pill',
-                  step === 4
+                  step === STEP_ORDER.length
                     ? 'bg-transparent'
-                    : step < currentStep
+                    : step < currentIndex
                       ? 'bg-charcoal-full'
                       : 'bg-slate-outline',
                 )}
