+++ nova-app/src/features/onboarding/ui/index.ts
@@ -1,6 +1,6 @@
-export { default as MajorCard } from './MajorCard';
-export { default as InterestCard } from './InterestCard';
-export { default as SkillCard } from './SkillCard';
-export { default as InterestKeywordCard } from './InterestKeywordCard';
+export { default as MajorCard } from './OnboardingContainer/MajorCard';
+export { default as InterestCard } from './OnboardingContainer/InterestCard';
+export { default as SkillCard } from './OnboardingContainer/SkillCard';
+export { default as InterestKeywordCard } from './OnboardingContainer/InterestKeywordCard';
 export { default as Stepper } from './Stepper';
 export { default as SearchInput } from './SearchInput';
